+++ simter-file-dao-jpa/src/test/kotlin/tech/simter/file/impl/dao/jpa/UpdateMethodImplTest.kt
@@ -7,9 +7,9 @@ import org.springframework.test.context.junit.jupiter.SpringJUnitConfig
 import reactor.kotlin.test.test
 import tech.simter.exception.NotFoundException
 import tech.simter.file.core.AttachmentDao
-import tech.simter.file.impl.dao.jpa.TestHelper.randomAttachmentId
 import tech.simter.file.impl.dao.jpa.po.AttachmentPo
 import tech.simter.file.impl.domain.AttachmentUpdateInfoImpl
+import tech.simter.file.test.TestHelper.randomAttachmentId
 import tech.simter.reactive.test.jpa.ReactiveDataJpaTest
 import tech.simter.reactive.test.jpa.TestEntityManager
 import java.time.OffsetDateTime