+++ simter-file-dao-mongo/src/test/kotlin/tech/simter/file/impl/dao/mongo/DeleteMethodImplTest.kt
@@ -7,8 +7,8 @@ import org.springframework.test.context.junit.jupiter.SpringJUnitConfig
 import reactor.kotlin.test.test
 import tech.simter.file.core.AttachmentDao
 import tech.simter.file.impl.dao.mongo.TestHelper.cleanDatabase
-import tech.simter.file.impl.dao.mongo.TestHelper.randomAttachmentId
 import tech.simter.file.impl.dao.mongo.TestHelper.randomAttachmentPo
+import tech.simter.file.test.TestHelper.randomAttachmentId
 
 /**
  * Test [AttachmentDaoImpl.delete].