+++ simter-file-dao-r2dbc/src/test/kotlin/tech/simter/file/impl/dao/r2dbc/GetFullPathMethodImplTest.kt
@@ -5,8 +5,8 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import reactor.kotlin.test.test
 import tech.simter.file.core.AttachmentDao
-import tech.simter.file.impl.dao.r2dbc.TestHelper.randomAttachmentId
 import tech.simter.file.impl.dao.r2dbc.po.AttachmentPo
+import tech.simter.file.test.TestHelper.randomAttachmentId
 import java.time.OffsetDateTime
 
 /**