+++ simter-file-rest-webflux/src/main/kotlin/tech/simter/file/rest/webflux/handler/DownloadFileHandler.kt
@@ -17,6 +17,7 @@ import org.springframework.web.reactive.function.server.ServerResponse.*
 import reactor.core.publisher.Mono
 import reactor.core.scheduler.Schedulers
 import tech.simter.exception.PermissionDeniedException
+import tech.simter.file.FILE_ROOT_DIR_KEY
 import tech.simter.file.core.AttachmentService
 import kotlin.text.Charsets.ISO_8859_1
 
@@ -60,7 +61,7 @@ import kotlin.text.Charsets.ISO_8859_1
  */
 @Component
 class DownloadFileHandler @Autowired constructor(
-  @Value("\${simter.file.root}") private val fileRootDir: String,
+  @Value("\${$FILE_ROOT_DIR_KEY}") private val fileRootDir: String,
   private val attachmentService: AttachmentService
 ) : HandlerFunction<ServerResponse> {
   override fun handle(request: ServerRequest): Mono<ServerResponse> {