+++ simter-file-rest-webflux/src/test/kotlin/tech/simter/file/rest/webflux/handler/UploadFileByFormHandlerTest.kt
@@ -19,6 +19,7 @@ import reactor.core.publisher.Mono
 import reactor.kotlin.test.test
 import tech.simter.exception.NotFoundException
 import tech.simter.exception.PermissionDeniedException
+import tech.simter.file.FILE_ROOT_DIR_KEY
 import tech.simter.file.core.AttachmentService
 import tech.simter.file.rest.webflux.UnitTestConfiguration
 import java.io.File
@@ -34,11 +35,11 @@ import java.util.*
 @SpringJUnitConfig(UnitTestConfiguration::class)
 @WebFluxTest
 @SpykBean(UploadFileByFormHandler::class)
-@TestPropertySource(properties = ["simter.file.root=target/files"])
+@TestPropertySource(properties = ["$FILE_ROOT_DIR_KEY=target/files"])
 class UploadFileByFormHandlerTest @Autowired constructor(
   private val client: WebTestClient,
   private val service: AttachmentService,
-  @Value("\${simter.file.root}") private val fileRootDir: String,
+  @Value("\${$FILE_ROOT_DIR_KEY}") private val fileRootDir: String,
   private val handler: UploadFileByFormHandler
 ) {
   @AfterEach
@@ -71,7 +72,7 @@ class UploadFileByFormHandlerTest @Autowired constructor(
       .contentLength(fileSize)
       .bodyValue(parts)
       .exchange()
-      .expectStatus().isNoContent
+      .expectStatus().isCreated
       .expectHeader().valueEquals("Location", "/$id")
 
     // verify